requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cached-property==2.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
fastapi==0.124.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
inflection==0.5.1
MarkupSafe==3.0.3
msal==1.34.0
mypy_extensions==1.1.0
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlalchemy-orm==1.2.10
starlette==0.50.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
Werkzeug==3.1.4
